Move `add_signatures` to `SpliceTransaction`
